requirements.txt
dask==0.12.0
decorator==4.0.10
networkx==1.11
numpy==1.11.2
Pillow==3.4.2
scikit-image==0.12.3
scikit-learn==0.18
scipy==0.18.1
six==1.10.0
toolz==0.8.0
